Extract `Client` into `email_shared`
