+++ email_broker/src/main.rs
@@ -2,26 +2,16 @@ mod config;
 
 use log::{error, info};
 use rusoto_core::RusotoError;
-use rusoto_dynamodb::{DynamoDb, DynamoDbClient, GetItemInput};
-use rusoto_sqs::{DeleteMessageBatchRequest, DeleteMessageBatchRequestEntry, Message};
-use rusoto_sqs::{ReceiveMessageError, ReceiveMessageRequest, Sqs, SqsClient};
+use rusoto_dynamodb::DynamoDbClient;
+use rusoto_sqs::{
+    DeleteMessageBatchRequest, DeleteMessageBatchRequestEntry, Message, ReceiveMessageError,
+    ReceiveMessageRequest, Sqs, SqsClient,
+};
 use simplelog::{Config as LogConfig, LevelFilter, TermLogger, TerminalMode};
-use std::convert::TryFrom;
 use structopt::StructOpt;
 
 use config::Options;
-use email_shared::email_message::{EmailMessage, ParseEmailMessageCode};
-use email_shared::queue::EmailPointerMessage;
-
-/// Hold references to external service clients so they only need to be allocated once.
-struct Client<'a> {
-    /// Connection to DynamoDB
-    dynamodb: &'a DynamoDbClient,
-    /// Connection to SQS
-    sqs: &'a SqsClient,
-    /// DynamoDB table from which email data will be read.
-    table_name: &'a str,
-}
+use email_shared::Client;
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
@@ -30,14 +20,10 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     info!("{:?}", opt);
     let sqs = SqsClient::new(opt.region.clone());
     let dynamodb = DynamoDbClient::new(opt.region.clone());
-    let client = Client {
-        dynamodb: &dynamodb,
-        sqs: &sqs,
-        table_name: &opt.table_name,
-    };
+    let client = Client::new(&dynamodb, &opt.table_name);
     let queue_url = &opt.queue_url;
     loop {
-        let message_list = get_sqs_email_messages(queue_url, client.sqs).await;
+        let message_list = get_sqs_email_messages(queue_url, &sqs).await;
         let processed_messages = match message_list {
             Ok(messages) => client.process_messages(messages).await,
             Err(error) => {
@@ -61,49 +47,6 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     Ok(())
 }
 
-impl Client<'_> {
-    async fn process_messages(&self, messages: Vec<Message>) -> Vec<EmailPointerMessage> {
-        info!("Process messages, {:?}", messages);
-        let mut processed_message_handles = Vec::new();
-        for message in messages {
-            match self.process_message(message).await {
-                Ok(id_message) => processed_message_handles.push(id_message),
-                Err(_) => (), // TODO: This needs to at least log the error
-            }
-        }
-        processed_message_handles
-    }
-
-    async fn process_message(&self, message: Message) -> Result<EmailPointerMessage, String> {
-        let pointer = EmailPointerMessage::try_from(message)?;
-        let email_message = self.get_email_message(&pointer).await;
-        let send_result = match email_message {
-            Ok(email) => send_email(email).await,
-            Err(error) => {
-                error!("process_message: {}: {}", &pointer, error);
-                Err("Unable to Parse Email".into())
-            }
-        };
-        match send_result {
-            Ok(_) => Ok(pointer),
-            Err(msg) => Err(msg),
-        }
-    }
-
-    async fn get_email_message(
-        &self,
-        message: &EmailPointerMessage,
-    ) -> Result<EmailMessage, ParseEmailMessageCode> {
-        let mut input = GetItemInput::from(message);
-        input.table_name = self.table_name.into();
-        self.dynamodb
-            .get_item(input)
-            .await
-            .map_err(ParseEmailMessageCode::from)
-            .and_then(EmailMessage::try_from)
-    }
-}
-
 /// Poll SQS at the given `queue_url` for new messages providing an iterator for `EmailIdMessage`.
 async fn get_sqs_email_messages(
     queue_url: &str,
@@ -120,9 +63,4 @@ async fn get_sqs_email_messages(
     sqs.receive_message(request)
         .await
         .map(|result| result.messages.unwrap_or(Vec::new()))
-}
-
-async fn send_email(email: EmailMessage) -> Result<(), String> {
-    info!("send_email: {:?}", email);
-    Err("Unimplemented".into())
 }
