+++ email_shared/src/lib.rs
@@ -1,7 +1,10 @@
 pub mod attribute_value_wrapper;
+mod client;
 pub mod dynamo;
 pub mod email_message;
 pub mod queue;
+
+pub use client::Client;
 
 #[cfg(test)]
 mod tests {
